+++ sdk/lib/_internal/pub/test/lish/package_creation_provides_a_malformed_success_test.dart
@@ -6,6 +6,7 @@ import 'dart:convert';
 
 import 'package:scheduled_test/scheduled_test.dart';
 import 'package:scheduled_test/scheduled_server.dart';
+import 'package:shelf/shelf.dart' as shelf;
 
 import '../descriptor.dart' as d;
 import '../test_pub.dart';
@@ -26,8 +27,7 @@ main() {
 
     var body = {'success': 'Your package was awesome.'};
     server.handle('GET', '/create', (request) {
-      request.response.write(JSON.encode(body));
-      request.response.close();
+      return new shelf.Response.ok(JSON.encode(body));
     });
 
     pub.stderr.expect('Invalid server response:');